♻️ Move breadcrumb text generation to VideoSettings.kt
Refactors the UI text generation for per-video breadcrumb settings from
RecentFilesAdapter to a `getBreadcrumbText` method on the VideoSettings
data class itself. This makes the logic reusable and cleans up the adapter.
